requirements.txt
Django>=4.0
djangorestframework
Pillow
python-decouple
psycopg2-binary
whitenoise
gunicorn
